Flask-based FullStack Employee Management System/requirements.txt
Flask>=3.0
mysql-connector-python>=8.3
python-dotenv>=1.0